requirements.txt
streamlit
python-dotenv
requests

# local packages
-e .